Feature: database h2

+++ dpelos/src/main/java/com/web/dpelos/service/DuenoService.java
@@ -5,13 +5,13 @@ import java.util.Collection;
 import com.web.dpelos.entity.Dueno;
 
 public interface DuenoService {
-    public Dueno buscarDuenoPorId(Integer id);
+    public Dueno buscarDuenoPorId(Long id);
 
     public Collection<Dueno> obtenerDuenos();
 
     public void addDueno(Dueno dueno);
 
-    public void deleteDueno(Integer id);
+    public void deleteDueno(Long id);
 
     public void updateDueno(Dueno dueno);
